Migrate to dask for efficient RAM Usage

+++ Week3/data/etl_web_to_gcs_bq.py
@@ -4,8 +4,10 @@ import glob
 import os
 import json
 import pandas as pd
+import dask.dataframe as dd
+from dask.distributed import Client
 import pandas_gbq
-from google.cloud import storage
+from google.cloud import storage, bigquery
 from google.oauth2.service_account import Credentials
 
 # %%
@@ -22,16 +24,17 @@ print(year_month_list)
 with open(PATH_DIRECTORY / "bq_config.json", "r") as f:
   bq_config = json.load(f)
 
-print(bq_config['project_id'])
-print(bq_config['credential_file'])
-
 credential = Credentials.from_service_account_file(bq_config['credential_file'])
+
+dask_client = Client(n_workers=1, threads_per_worker=4, processes=False)
 
 # %%
 def load_files() -> dict:
   dataframe_collection = {}
   for month in year_month_list:
-    dataframe_collection[month] = pd.DataFrame(pd.read_csv(f"{PATH_DIRECTORY}/fhv_tripdata_{month}.csv.gz"))
+    if not os.path.isfile(f"{PATH_DIRECTORY}/fhv_tripdata_{month}.csv.gz"):
+      os.system(f"wget https://github.com/DataTalksClub/nyc-tlc-data/releases/download/fhv/fhv_tripdata_{month}.csv.gz -P {PATH_DIRECTORY}/")
+    dataframe_collection[month] = dd.DataFrame(dd.read_csv(f"{PATH_DIRECTORY}/fhv_tripdata_{month}.csv.gz"))
   
   return dataframe_collection
 
@@ -47,28 +50,26 @@ def print_dataframe(collection: dict) -> None:
 def transform(collection: dict) -> dict:
   for key in collection.keys():
     df = collection[key]
-    df["pickup_datetime"] = pd.to_datetime(df["pickup_datetime"])
-    df["dropOff_datetime"] = pd.to_datetime(df["dropOff_datetime"])
-    # print(f"pre: missing PUlocationID {key} count: {df['PUlocationID'].isna().sum()}")
-    # df["PUlocationID"].fillna(0, inplace=True) 
-    # print(f"post: missing PUlocationID {key} count: {df['PUlocationID'].isna().sum()}")
+    df["pickup_datetime"] = dd.to_datetime(df["pickup_datetime"])
+    df["dropOff_datetime"] = dd.to_datetime(df["dropOff_datetime"])
     collection[key] = df
   return collection
 
 # %%
 def write_local(collection: dict) -> None:
   for month in year_month_list:
-    df: pd.DataFrame = collection[month]
-    # collection[month] = pd.DataFrame(pd.read_csv(f"{PATH_DIRECTORY}/fhv_tripdata_{month}.csv.gz"))
-    df.to_parquet(f"{PATH_DIRECTORY}/parquet/fhv_tripdata_{month}.parquet")
+    if not os.path.isfile(f"{PATH_DIRECTORY}/parquet/fhv_tripdata_{month}.parquet"):
+      df: dd.DataFrame = collection[month]
+      df.to_parquet(f"{PATH_DIRECTORY}/parquet/fhv_tripdata_{month}.parquet")
 
 def write_gcs() -> None:
   client = storage.Client(credentials=credential, project=bq_config['project_id'])
-  bucket = storage.Bucket(client, 'prefect-de-zoomcamp')
+  bucket = storage.Bucket(client, bq_config["bucket_id"])
   files = glob.glob(f"{PATH_DIRECTORY}/parquet/*.parquet")
   for file in files:
     filename = os.path.basename(file)
-    blob = bucket.blob(f"data/{filename}")
+    blob = bucket.blob(f"data/{filename}", chunk_size=8388608)
+    print(f"upload {filename}")
     blob.upload_from_filename(file)
     print(f"{filename} uploaded")
 
@@ -79,18 +80,46 @@ def web_to_bq(collection: dict) -> None:
     pandas_gbq.to_gbq(
       collection[key], 
       "trips_data_all.data_trip", 
-      project_id="atlantean-glyph-376012", 
+      project_id=bq_config["project_id"], 
       progress_bar=True,
       chunksize='500_000',
       if_exists='append', 
       credentials=credential
     )
+
+# %%
+def create_external_table() -> None:
+  client = bigquery.Client(credentials=credential,project=bq_config['project_id'])
+  dataset_id = "trips_data_all"
+  table_id = "fvh_trip"
+  bucket_id = bq_config['bucket_id']
+
+  job_config = bigquery.QueryJobConfig()
+  table_ref = client.dataset(dataset_id).table(table_id)
+  job_config.destination = table_ref
+  sql = f"""
+         CREATE OR REPLACE EXTERNAL TABLE `{bq_config['project_id']}.{dataset_id}.{table_id}`
+         OPTIONS (
+            format = 'PARQUET',
+            uris = ['gs://{bucket_id}/data/fvh_tripdata_2019-*.parquet]
+         );
+         """
+  
+  query_job = client.query(
+    sql,
+    job_config=job_config,
+  )
+
+  query_job.result()
+  print(f"Query results loaded to table {table_ref.path}")
+
+  dask_client.close()
   
 # %%
 if __name__ == '__main__':
-  # dataframe_collection = load_files()
-#   # print_dataframe(dataframe_collection)
-  # transform_collection = transform(dataframe_collection)
-  # write_local(transform_collection)
+  dataframe_collection = load_files()
+  transform_collection = transform(dataframe_collection)
+  write_local(transform_collection)
   write_gcs()
-#   web_to_bq(transform_collection)
+  web_to_bq(transform_collection)
+  create_external_table()
